// src/types/schemas/session.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface SessionSchema {
  sessionDatetime: string;
  sessionName: string;
  sessionType: "Practice" | "Qualifying" | "Race" | "Testing" | "Other";
}